move commands to separate folders
