+++ lib/commands/init/make-init.js
@@ -1,8 +1,9 @@
-const { makeCfService } = require('./cf/make-cf-service');
 const get = require('lodash.get');
 const Bluebird = require('bluebird');
 const path = require('path');
-const { log } = require('../utils/logger');
+
+const { log } = require('../../utils/logger');
+const { makeCfService } = require('../../services//cf/make-cf-service');
 
 const initSsm = () => Bluebird.resolve();
