+++ lib/commands/list/make-list.js
@@ -2,8 +2,9 @@
 
 const Bluebird = require('bluebird');
 const chalk = require('chalk');
-const { log } = require('../utils/logger');
 const get = require('lodash.get');
+
+const { log } = require('../../utils/logger');
 
 const getParameters = ({ parameterNames, parameterStore }) => {
   if(!parameterNames.length) {
