+++ lib/commands/run/populate-config.js
@@ -1,13 +1,12 @@
 'use strict';
 
 const Bluebird = require('bluebird');
-
-const { promptRequiredConfigs } = require('../prompt-required-configs');
 const isEmpty = require('lodash.isempty');
 const get = require('lodash.get');
-
 const chalk = require('chalk');
-const { log } = require('../utils/logger');
+
+const { log } = require('../../utils/logger');
+const { promptRequiredConfigs } = require('./prompt-required-configs');
 
 const populateConfig = ({ settings, config, overrides, interactive, parameterStore }) => {
   log(chalk.white('Populating configuration in SSM...'));
