+++ lib/commands/run/populate-secret.js
@@ -1,11 +1,11 @@
 'use strict';
 
 const Bluebird = require('bluebird');
-
-const { promptRequiredConfigs } = require('../prompt-required-configs');
 const chalk = require('chalk');
 const get = require('lodash.get');
-const { log } = require('../utils/logger');
+
+const { log } = require('../../utils/logger');
+const { promptRequiredConfigs } = require('./prompt-required-configs');
 
 const populateSecret = ({ settings, secret, interactive, parameterStore }) => {
   log(chalk.white('Populating secrets in SSM...'));
