+++ lib/commands/run/prompt-required-configs.js
@@ -6,7 +6,7 @@ const get = require('lodash.get');
 const chalk = require('chalk');
 const prompt = Bluebird.promisifyAll(require('prompt'));
 
-const { logWarning } = require('./utils/logger');
+const { logWarning } = require('../../utils/logger');
 
 const getAllExistingConfigs = (requiredConfigs, currentConfigs) => {
   if (!requiredConfigs) {
