+++ lib/makeOprah.js
@@ -4,9 +4,9 @@ const chalk = require('chalk');
 const { log } = require('../lib/utils/logger');
 const { makeParameterStore } = require('./services/parameter-store/make-parameter-store');
 const { makeSettingsService } = require('./services/settings/make-settings-service');
-const { makeList } = require('./services/make-list');
-const { makeInit } = require('./services/make-init');
-const { makeRun } = require('./services/make-run');
+const { makeList } = require('./commands/list/make-list');
+const { makeInit } = require('./commands/init/make-init');
+const { makeRun } = require('./commands/run/make-run');
 
 const makeOprah = ({ params }) => {
   const settingsService = makeSettingsService({
